feat(msm8916): implement basic CPU power management
Implement very basic support for CPU power management that allows
turning off or suspending CPU cores. Cluster power management needs
more investigation and might be added later.

The main relevant hardware block for this is the SAW/SPM that needs
some configuration and is then set up to go to different power states
when the CPU executes WFI. As far as I can tell there is no separate
way to turn a CPU off and back on (for PSCI CPU_OFF/ON). However, the
CPU can be configured to go to suspend ("power collapse") and then
woken up again by sending a SGI via the GICv2.

Change-Id: If4cbc68d56ca2f0e367a876b72b396545d0c1314
Signed-off-by: Stephan Gerhold <stephan@gerhold.net>